# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MIDI ML Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Genre {
  id: string;
  name: string;
  description: string;
}

export type JobType = (typeof JobType)[keyof typeof JobType];

export const JobType = {
  extend: "extend",
  transform: "transform",
  evaluate: "evaluate",
  live_extend: "live_extend",
} as const;

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface MidiFeatures {
  totalNotes?: number;
  uniquePitches?: number;
  estimatedTempo?: number;
  trackCount?: number;
  durationSeconds?: number;
}

export interface EvaluationResult {
  overallScore?: number;
  rhythmScore?: number;
  harmonyScore?: number;
  melodyScore?: number;
  complexityScore?: number;
  predictedGenre?: string;
  genreConfidence?: number;
  midiFeatures?: MidiFeatures;
  lecturerFeedback?: string;
  summary?: string;
  suggestions?: string[];
}

export interface Job {
  id: number;
  type: JobType;
  status: JobStatus;
  inputFilename: string;
  /** @nullable */
  outputFilename?: string | null;
  /** @nullable */
  targetGenre?: string | null;
  /** @nullable */
  barsToExtend?: number | null;
  evaluationResult?: EvaluationResult;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ExtendInput {
  inputFilename: string;
  /**
   * @minimum 1
   * @maximum 64
   */
  barsToExtend: number;
}

export interface TransformInput {
  inputFilename: string;
  targetGenre: string;
}

export interface EvaluateInput {
  inputFilename: string;
}

export interface LiveExtendInput {
  inputFilename: string;
  /**
   * @minimum 1
   * @maximum 8
   */
  barsToExtend?: number;
}

export interface UploadResult {
  filename: string;
  originalName: string;
  size: number;
}

export interface DownloadInfo {
  jobId: number;
  filename: string;
  url: string;
}

export interface StatsByType {
  extend: number;
  transform: number;
  evaluate: number;
  live_extend: number;
}

export interface Stats {
  totalJobs: number;
  completedJobs: number;
  failedJobs: number;
  byType: StatsByType;
}

export type UploadMidiFileBody = {
  file: Blob;
};