# lib/api-zod/src/generated/types/evaluationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MIDI ML Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { MidiFeatures } from "./midiFeatures";

export interface EvaluationResult {
  overallScore?: number;
  rhythmScore?: number;
  harmonyScore?: number;
  melodyScore?: number;
  complexityScore?: number;
  predictedGenre?: string;
  genreConfidence?: number;
  midiFeatures?: MidiFeatures;
  lecturerFeedback?: string;
  summary?: string;
  suggestions?: string[];
}