# lib/api-zod/src/generated/types/genre.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MIDI ML Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface Genre {
  id: string;
  name: string;
  description: string;
}