# lib/api-zod/src/generated/types/liveExtendInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MIDI ML Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface LiveExtendInput {
  inputFilename: string;
  /**
   * @minimum 1
   * @maximum 8
   */
  barsToExtend?: number;
}