# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MIDI ML Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsByType } from "./statsByType";

export interface Stats {
  totalJobs: number;
  completedJobs: number;
  failedJobs: number;
  byType: StatsByType;
}