# lib/api-zod/src/generated/types/statsByType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MIDI ML Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsByType {
  extend: number;
  transform: number;
  evaluate: number;
  live_extend: number;
}